requirements.txt
jsonschema>=4.0
drf_spectacular
pytest
pytest-django